+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### What's new
 - Customer, Subscriber and Guest syncs have been refactored for speed and efficiency.
 
+### Bug fixes
+- We've replaced the requirement for Symfony\Component\HttpFoundation in our Server Sent Events helper.
+
 # 4.18.2
 
 ### Improvements